split archive command logic into separate module and update tests
