+++ src/steps/copy/index.js
@@ -3,7 +3,8 @@ import path from 'path'
 import { fileURLToPath } from 'url'
 import { copyContent } from '../../utils/copy.js'
 import { error, header, success } from '../../utils/exec.js'
-import { patchAgentGuidance, patchAgentsMd, patchArchiveCommand, patchConcurrency } from './agents.js'
+import { patchAgentGuidance, patchAgentsMd, patchConcurrency } from './agents.js'
+import { patchArchiveCommand } from './commands.js'
 import { installSkills } from './skills.js'
 import { exit } from '../../utils/process.js'
